AI/requirements.txt
livekit
livekit-agents[deepgram,silero]
python-dotenv
aiohttp
ollama
